Add --vmaf ARG
