+++ src/vmaf.rs
@@ -3,7 +3,7 @@ use crate::{
     process::{exit_ok, CommandExt, FfmpegProgress},
     yuv,
 };
-use anyhow::{ensure, Context};
+use anyhow::Context;
 use std::path::Path;
 use tokio::process::Command;
 use tokio_process_stream::{Item, ProcessChunkStream};
@@ -14,27 +14,16 @@ use tokio_stream::{Stream, StreamExt};
 pub fn run(
     reference: &Path,
     distorted: &Path,
-    options: Option<&str>,
+    lavfi: &str,
 ) -> anyhow::Result<impl Stream<Item = VmafOut>> {
     let (yuv_out, yuv_pipe) = yuv::pipe420p10le(reference)?;
     let yuv_pipe = yuv_pipe.filter_map(VmafOut::ignore_ok);
-
-    let libvmaf_options = match options {
-        None => "libvmaf".into(),
-        Some(opts) => {
-            ensure!(
-                !opts.contains('\''),
-                "invalid vmaf-options: must not contain `'` character"
-            );
-            format!("libvmaf='{opts}'")
-        }
-    };
 
     let vmaf: ProcessChunkStream = Command::new("ffmpeg")
         .kill_on_drop(true)
         .arg2("-i", distorted)
         .arg2("-i", "-")
-        .arg2("-lavfi", &libvmaf_options)
+        .arg2("-lavfi", lavfi)
         .arg2("-f", "null")
         .arg("-")
         .stdin(yuv_out)
